+++ console/src/main/java/org/nervos/ckb/example/transaction/CollectUtils.java
@@ -3,16 +3,15 @@ package org.nervos.ckb.example.transaction;
 import java.io.IOException;
 import java.math.BigInteger;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import org.nervos.ckb.address.AddressUtils;
-import org.nervos.ckb.methods.type.Script;
-import org.nervos.ckb.methods.type.cell.CellOutput;
-import org.nervos.ckb.service.CKBService;
+import org.nervos.ckb.service.Api;
 import org.nervos.ckb.system.type.SystemScriptCell;
-import org.nervos.ckb.transaction.CellGatherer;
+import org.nervos.ckb.transaction.CellCollector;
 import org.nervos.ckb.transaction.CollectedCells;
 import org.nervos.ckb.transaction.Utils;
+import org.nervos.ckb.type.Script;
+import org.nervos.ckb.type.cell.CellOutput;
 import org.nervos.ckb.utils.Network;
 import org.nervos.ckb.utils.Numeric;
 
@@ -20,13 +19,13 @@ import org.nervos.ckb.utils.Numeric;
 public class CollectUtils {
 
   private SystemScriptCell systemSecpCell;
-  private CKBService ckbService;
+  private Api api;
   private BigInteger collectedCapacity = BigInteger.ZERO;
 
-  public CollectUtils(CKBService ckbService) {
-    this.ckbService = ckbService;
+  public CollectUtils(Api api) {
+    this.api = api;
     try {
-      systemSecpCell = Utils.getSystemScriptCell(ckbService);
+      systemSecpCell = Utils.getSystemScriptCell(api);
     } catch (IOException e) {
       e.printStackTrace();
     }
@@ -39,9 +38,9 @@ public class CollectUtils {
           Utils.generateLockScriptWithPrivateKey(sender.privateKey, systemSecpCell.cellHash)
               .computeHash();
       CollectedCells collectedCells =
-          new CellGatherer(ckbService).getCellInputs(lockHash, sender.capacity);
+          new CellCollector(api).getCellInputs(lockHash, sender.capacity);
       if (collectedCells.capacity.compareTo(sender.capacity) < 0) {
-        throw new IOException("No enough Capacities with sender private key: " + sender.privateKey);
+        throw new IOException("No enough Capacity with sender private key: " + sender.privateKey);
       }
       collectedCapacity = collectedCapacity.add(collectedCells.capacity);
       cellsWithPrivateKeys.add(new CellsWithPrivateKey(collectedCells.inputs, sender.privateKey));
@@ -57,8 +56,7 @@ public class CollectUtils {
       cellOutputs.add(
           new CellOutput(
               receiver.capacity.toString(),
-              new Script(
-                  systemSecpCell.cellHash, Collections.singletonList(blake160), Script.TYPE)));
+              new Script(systemSecpCell.cellHash, blake160, Script.TYPE)));
     }
     BigInteger needCapacity = BigInteger.ZERO;
     for (Receiver receiver : receivers) {
@@ -71,7 +69,7 @@ public class CollectUtils {
               collectedCapacity.subtract(needCapacity).toString(),
               new Script(
                   systemSecpCell.cellHash,
-                  Collections.singletonList(Numeric.prependHexPrefix(changeAddressBlake160)),
+                  Numeric.prependHexPrefix(changeAddressBlake160),
                   Script.TYPE)));
     }
     return cellOutputs;