+++ core/src/main/java/org/nervos/ckb/exceptions/InvalidNumberOfWitnessesException.java
@@ -1,9 +1,0 @@
-package org.nervos.ckb.exceptions;
-
-/** Copyright © 2019 Nervos Foundation. All rights reserved. */
-public class InvalidNumberOfWitnessesException extends RuntimeException {
-
-  public InvalidNumberOfWitnessesException(String message) {
-    super(message);
-  }
-}