+++ core/src/main/java/org/nervos/ckb/methods/type/Cycles.java
@@ -1,7 +1,0 @@
-package org.nervos.ckb.methods.type;
-
-/** Copyright © 2019 Nervos Foundation. All rights reserved. */
-public class Cycles {
-
-  public String clcles;
-}