+++ core/src/main/java/org/nervos/ckb/transaction/TransactionBuilder.java
@@ -8,14 +8,12 @@ import java.util.List;
 import org.nervos.ckb.crypto.Blake2b;
 import org.nervos.ckb.crypto.secp256k1.ECKeyPair;
 import org.nervos.ckb.crypto.secp256k1.Sign;
-import org.nervos.ckb.exceptions.InvalidNumberOfWitnessesException;
-import org.nervos.ckb.methods.type.Witness;
-import org.nervos.ckb.methods.type.cell.CellDep;
-import org.nervos.ckb.methods.type.cell.CellInput;
-import org.nervos.ckb.methods.type.cell.CellOutput;
-import org.nervos.ckb.methods.type.transaction.Transaction;
-import org.nervos.ckb.service.CKBService;
+import org.nervos.ckb.service.Api;
 import org.nervos.ckb.system.type.SystemScriptCell;
+import org.nervos.ckb.type.cell.CellDep;
+import org.nervos.ckb.type.cell.CellInput;
+import org.nervos.ckb.type.cell.CellOutput;
+import org.nervos.ckb.type.transaction.Transaction;
 import org.nervos.ckb.utils.Numeric;
 
 /** Copyright © 2019 Nervos Foundation. All rights reserved. */
@@ -27,12 +25,12 @@ public class TransactionBuilder {
   private List<CellInput> cellInputs = new ArrayList<>();
   private List<CellOutput> cellOutputs = new ArrayList<>();
   private List<String> cellOutputsData = new ArrayList<>();
-  private List<Witness> witnesses = new ArrayList<>();
+  private List<String> witnesses = new ArrayList<>();
   private Transaction transaction;
 
-  public TransactionBuilder(CKBService ckbService) {
+  public TransactionBuilder(Api api) {
     try {
-      this.systemSecpCell = Utils.getSystemScriptCell(ckbService);
+      this.systemSecpCell = Utils.getSystemScriptCell(api);
     } catch (Exception e) {
       e.printStackTrace();
     }
@@ -67,7 +65,7 @@ public class TransactionBuilder {
     }
     for (int i = 0; i < cellOutputs.size(); i++) {
       cellOutputsData.add("0x");
-      witnesses.add(new Witness());
+      witnesses.add("0x");
     }
     transaction =
         new Transaction(
@@ -85,34 +83,27 @@ public class TransactionBuilder {
       throw new IOException("Transaction could not null");
     }
     if (witnesses.size() < cellInputs.size()) {
-      throw new InvalidNumberOfWitnessesException("Invalid number of witnesses");
+      throw new IOException("Invalid number of witnesses");
     }
     witnesses.set(index, signWitness(witnesses.get(index), privateKey));
   }
 
   public void sign(String privateKey) {
-    for (Witness witness : witnesses) {
+    for (String witness : witnesses) {
       witnesses.add(signWitness(witness, privateKey));
     }
   }
 
-  private Witness signWitness(Witness witness, String privateKey) {
+  private String signWitness(String witness, String privateKey) {
     ECKeyPair ecKeyPair = ECKeyPair.createWithPrivateKey(privateKey, false);
-    List<String> oldData = witness.data;
     Blake2b blake2b = new Blake2b();
     blake2b.update(Numeric.hexStringToByteArray(transaction.computeHash()));
-    for (String datum : witness.data) {
-      blake2b.update(Numeric.hexStringToByteArray(datum));
-    }
+    blake2b.update(Numeric.hexStringToByteArray(witness));
     String message = blake2b.doFinalString();
-
     String signature =
         Numeric.toHexString(
             Sign.signMessage(Numeric.hexStringToByteArray(message), ecKeyPair).getSignature());
-    witness.data = new ArrayList<>();
-    witness.data.add(signature);
-    witness.data.addAll(oldData);
-    return witness;
+    return signature + Numeric.cleanHexPrefix(witness);
   }
 
   public Transaction getTransaction() {