+++ core/src/main/java/org/nervos/ckb/type/Block.java
@@ -1,0 +1,22 @@
+package org.nervos.ckb.type;
+
+import java.util.List;
+import org.nervos.ckb.type.transaction.Transaction;
+
+/** Copyright © 2018 Nervos Foundation. All rights reserved. */
+public class Block {
+  public Header header;
+
+  public List<Transaction> transactions;
+
+  public List<String> proposals;
+
+  public List<Uncle> uncles;
+
+  public static class Uncle {
+
+    public Header header;
+
+    public List<String> proposals;
+  }
+}