+++ serialization/src/main/java/org/nervos/ckb/type/dynamic/Dynamic.java
@@ -1,12 +1,11 @@
 package org.nervos.ckb.type.dynamic;
 
 import java.util.List;
-import org.nervos.ckb.type.base.DynType;
 import org.nervos.ckb.type.base.Type;
 import org.nervos.ckb.type.fixed.UInt32;
 
 /** Copyright © 2019 Nervos Foundation. All rights reserved. */
-public class Dynamic<T extends DynType> implements Type<List<T>> {
+public class Dynamic<T extends Type> implements Type<List<T>> {
   private List<T> value;
 
   public Dynamic(List<T> value) {
@@ -25,7 +24,7 @@ public class Dynamic<T extends DynType> implements Type<List<T>> {
     int offset = UInt32.BYTE_SIZE;
     int bytesOffset = UInt32.BYTE_SIZE * (1 + value.size());
 
-    for (DynType type : value) {
+    for (Type type : value) {
       // offset of every Bytes
       byte[] offsetBytes = new UInt32(bytesOffset).toBytes();
       System.arraycopy(offsetBytes, 0, dest, offset, UInt32.BYTE_SIZE);
@@ -47,7 +46,7 @@ public class Dynamic<T extends DynType> implements Type<List<T>> {
   @Override
   public int getLength() {
     int length = (1 + value.size()) * UInt32.BYTE_SIZE;
-    for (DynType type : value) {
+    for (Type type : value) {
       length += type.getLength();
     }
     return length;